requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
bs4==0.0.1
certifi==2020.6.20
chardet==3.0.4
dj-database-url==0.5.0
Django==3.1.1
django-filter==2.3.0
djangorestframework==3.11.1
idna==2.10
isort==5.5.3
lazy-object-proxy==1.4.3
mccabe==0.6.1

pylint==2.6.0
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.10
whitenoise==5.2.0
wrapt==1.12.1
